test(194): split remove-worktrees tests to satisfy 500-line file limit
Split test/remove-worktrees.test.ts (612 lines) along the production-module
boundary so each test file is under the repository's 500-line limit:
- test/remove-worktrees.test.ts (336 lines) retains the pure-logic tests.
- test/remove-worktrees-runner.test.ts (278 lines, new) holds the GitRunner
  and removeAllSecondaryWorktrees orchestration tests.

All 28 worktree tests preserved (388 total); no production code changed.
